% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_assignment.R
\name{question}
\alias{question}
\title{Add a question to the qmd file}
\usage{
question(insert = TRUE)
}
\arguments{
\item{insert}{whether to insert or print the question}
}
\description{
Add a question to the qmd file
}
