% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{render}
\alias{render}
\title{Render assignment file}
\usage{
render(questions = TRUE, solutions = TRUE, append_hash = TRUE)
}
\arguments{
\item{questions}{\if{html}{\out{<bool>}} whether to render the questions}

\item{append_hash}{\if{html}{\out{<bool>}} whether to append a hash to the solutions file.
See details.}

\item{answers}{\if{html}{\out{<bool>}} whether to render the solutions}
}
\description{
This function renders an assignment .qmd to html in two ways:
one with only the questions, or/and one which also includes the
solutions.
}
\details{
The hash can be useful if you host both the questions and
solutions on a single webserver, and you want the students
to not be able to view the answers before you choose to make
them available via a link.
}
